+++ ingeniamotion/pdo.py
@@ -2,14 +2,26 @@ import threading
 import time
 from collections import deque
 from copy import deepcopy
-from typing import TYPE_CHECKING, Callable, Deque, Dict, List, Optional, Tuple, Type, Union
+from typing import (
+    TYPE_CHECKING,
+    Callable,
+    Deque,
+    Dict,
+    List,
+    Optional,
+    Tuple,
+    Type,
+    Union,
+)
 
 from ingenialink.canopen.network import CanopenNetwork
+from ingenialink.enums.register import RegCyclicType
 from ingenialink.ethercat.network import EthercatNetwork
 from ingenialink.ethercat.register import EthercatRegister
 from ingenialink.ethercat.servo import EthercatServo
-from ingenialink.exceptions import ILError, ILStateError, ILWrongWorkingCount
+from ingenialink.exceptions import ILError, ILWrongWorkingCount
 from ingenialink.pdo import RPDOMap, RPDOMapItem, TPDOMap, TPDOMapItem
+
 from ingeniamotion.enums import COMMUNICATION_TYPE
 from ingeniamotion.exceptions import IMException
 from ingeniamotion.metaclass import DEFAULT_AXIS, DEFAULT_SERVO
@@ -26,6 +38,7 @@ class PDOPoller:
         mc: "MotionController",
         servo: str,
         refresh_time: float,
+        watchdog_timeout: Optional[float],
         buffer_size: int,
     ) -> None:
         """Constructor.
@@ -34,6 +47,8 @@ class PDOPoller:
             mc: MotionController instance
             servo: drive alias.
             refresh_time: PDO values refresh time.
+            watchdog_timeout: The PDO watchdog time. If not provided it will be set proportional
+             to the refresh rate.
             buffer_size: Maximum number of register readings to store.
 
         """
@@ -41,6 +56,7 @@ class PDOPoller:
         self.__mc = mc
         self.__servo = servo
         self.__refresh_time = refresh_time
+        self.__watchdog_timeout = watchdog_timeout
         self.__buffer_size = buffer_size
         self.__buffer: List[Deque[Union[int, float]]] = []
         self.__timestamps: Deque[float] = deque(maxlen=self.__buffer_size)
@@ -59,7 +75,9 @@ class PDOPoller:
         for callback in self.__exception_callbacks:
             self.__mc.capture.pdo.subscribe_to_exceptions(callback)
         self.__start_time = time.time()
-        self.__mc.capture.pdo.start_pdos(refresh_rate=self.__refresh_time)
+        self.__mc.capture.pdo.start_pdos(
+            refresh_rate=self.__refresh_time, watchdog_timeout=self.__watchdog_timeout
+        )
 
     def stop(self) -> None:
         """Stop the poller"""
@@ -170,21 +188,31 @@ class PDONetworkManager:
         Args:
             net: The EthercatNetwork instance where the PDOs will be active.
             refresh_rate: Determines how often (seconds) the PDO values will be updated.
+            watchdog_timeout: The PDO watchdog time. If not provided it will be set proportional
+             to the refresh rate.
             notify_send_process_data: Callback to notify when process data is about to be sent.
             notify_receive_process_data: Callback to notify when process data is received.
             notify_exceptions: Callback to notify when an exception is raised.
+
+        Raises:
+            ValueError: If the provided refresh rate is unfeasible.
+
         """
 
-        DEFAULT_PDO_REFRESH_RATE = 0.01
-        MINIMUM_PDO_REFRESH_RATE = 4
-        ETHERCAT_PDO_WATCHDOG = "processdata"
-        PDO_WATCHDOG_INCREMENT_FACTOR = 1.5
-        SECONDS_TO_MS_CONVERSION_FACTOR = 1000
+        DEFAULT_PDO_REFRESH_TIME = 0.01
+        MINIMUM_PDO_REFRESH_TIME = 0.001
+        MAXIMUM_PDO_REFRESH_TIME = 4
+        DEFAULT_WATCHDOG_TIMEOUT = 0.1
+        PDO_WATCHDOG_INCREMENT_FACTOR = 2
+        # The time.sleep precision is 13 ms for Windows OS
+        # https://stackoverflow.com/questions/1133857/how-accurate-is-pythons-time-sleep
+        WINDOWS_TIME_SLEEP_PRECISION = 0.013
 
         def __init__(
             self,
             net: EthercatNetwork,
             refresh_rate: Optional[float],
+            watchdog_timeout: Optional[float],
             notify_send_process_data: Optional[Callable[[], None]] = None,
             notify_receive_process_data: Optional[Callable[[], None]] = None,
             notify_exceptions: Optional[Callable[[IMException], None]] = None,
@@ -192,23 +220,28 @@ class PDONetworkManager:
             super().__init__()
             self._net = net
             if refresh_rate is None:
-                refresh_rate = self.DEFAULT_PDO_REFRESH_RATE
-            elif refresh_rate > self.MINIMUM_PDO_REFRESH_RATE:
+                refresh_rate = self.DEFAULT_PDO_REFRESH_TIME
+            elif refresh_rate < self.MINIMUM_PDO_REFRESH_TIME:
                 raise ValueError(
-                    f"The minimum PDO refresh rate is {self.MINIMUM_PDO_REFRESH_RATE} seconds."
+                    f"The minimum PDO refresh rate is {self.MINIMUM_PDO_REFRESH_TIME} seconds."
+                )
+            elif refresh_rate > self.MAXIMUM_PDO_REFRESH_TIME:
+                raise ValueError(
+                    f"The maximum PDO refresh rate is {self.MAXIMUM_PDO_REFRESH_TIME} seconds."
                 )
             self._refresh_rate = refresh_rate
-            self._pd_thread_stop_event = threading.Event()
-            for servo in self._net.servos:
-                servo.slave.set_watchdog(
-                    self.ETHERCAT_PDO_WATCHDOG,
-                    self._refresh_rate
-                    * self.PDO_WATCHDOG_INCREMENT_FACTOR
-                    * self.SECONDS_TO_MS_CONVERSION_FACTOR,
+            if watchdog_timeout is None:
+                watchdog_timeout = max(
+                    self.DEFAULT_WATCHDOG_TIMEOUT,
+                    self._refresh_rate * self.PDO_WATCHDOG_INCREMENT_FACTOR,
                 )
+            self._watchdog_timeout = watchdog_timeout
+            for servo in self._net.servos:
+                servo.set_pdo_watchdog_time(watchdog_timeout)
             self._notify_send_process_data = notify_send_process_data
             self._notify_receive_process_data = notify_receive_process_data
             self._notify_exceptions = notify_exceptions
+            self._pd_thread_stop_event = threading.Event()
 
         def run(self) -> None:
             """Start the PDO exchange"""
@@ -223,22 +256,22 @@ class PDONetworkManager:
                     self._notify_exceptions(im_exception)
             iteration_duration: float = -1
             while not self._pd_thread_stop_event.is_set():
-                time_start = time.time()
+                time_start = time.perf_counter()
                 if self._notify_send_process_data is not None:
                     self._notify_send_process_data()
                 try:
-                    self._net.send_receive_processdata()
+                    self._net.send_receive_processdata(self._refresh_rate)
                 except ILWrongWorkingCount as il_error:
                     self._pd_thread_stop_event.set()
                     self._net.stop_pdos()
                     if iteration_duration == -1:
-                        iteration_duration = time.time() - time_start
+                        iteration_duration = time.perf_counter() - time_start
                     duration_error = ""
-                    if iteration_duration > self._refresh_rate:
+                    if iteration_duration > self._watchdog_timeout:
                         duration_error = (
                             f"Last iteration took {iteration_duration * 1000:0.1f} ms which is higher"
-                            f" than the refresh rate ({self._refresh_rate * 1000:0.1f} ms). Please"
-                            " optimize the callbacks and/or increase the refresh rate."
+                            f" than the watchdog timeout ({self._watchdog_timeout * 1000:0.1f} ms). Please"
+                            " optimize the callbacks and/or increase the refresh rate/watchdog timeout."
                         )
                     if self._notify_exceptions is not None:
                         im_exception = IMException(
@@ -249,16 +282,29 @@ class PDONetworkManager:
                 else:
                     if self._notify_receive_process_data is not None:
                         self._notify_receive_process_data()
-                    remaining_loop_time = self._refresh_rate - (time.time() - time_start)
-                    if remaining_loop_time > 0:
-                        time.sleep(remaining_loop_time)
-                    iteration_duration = time.time() - time_start
+                    while (
+                        remaining_loop_time := self._refresh_rate
+                        - (time.perf_counter() - time_start)
+                    ) > 0:
+                        if remaining_loop_time > self.WINDOWS_TIME_SLEEP_PRECISION:
+                            time.sleep(self.WINDOWS_TIME_SLEEP_PRECISION)
+                        else:
+                            self.high_precision_sleep(remaining_loop_time)
+                    iteration_duration = time.perf_counter() - time_start
 
         def stop(self) -> None:
             """Stop the PDO exchange"""
             self._pd_thread_stop_event.set()
             self._net.stop_pdos()
             self.join()
+
+        @staticmethod
+        def high_precision_sleep(duration: float) -> None:
+            """Replaces the time.sleep() method in order to obtain
+            more precise sleeping times."""
+            start_time = time.perf_counter()
+            while duration - (time.perf_counter() - start_time) > 0:
+                pass
 
     def __init__(self, motion_controller: "MotionController") -> None:
         self.mc = motion_controller
@@ -291,9 +337,9 @@ class PDONetworkManager:
             AttributeError: If an initial value is not provided for an RPDO register.
 
         """
-        pdo_map_item_dict: Dict[str, Type[Union[RPDOMapItem, TPDOMapItem]]] = {
-            "CYCLIC_RX": RPDOMapItem,
-            "CYCLIC_TX": TPDOMapItem,
+        pdo_map_item_dict: Dict[RegCyclicType, Type[Union[RPDOMapItem, TPDOMapItem]]] = {
+            RegCyclicType.RX: RPDOMapItem,
+            RegCyclicType.TX: TPDOMapItem,
         }
         drive = self.mc._get_drive(servo)
         register = drive.dictionary.registers(axis)[register_uid]
@@ -481,6 +527,7 @@ class PDONetworkManager:
         self,
         network_type: Optional[COMMUNICATION_TYPE] = None,
         refresh_rate: Optional[float] = None,
+        watchdog_timeout: Optional[float] = None,
     ) -> None:
         """
         Start the PDO exchange process.
@@ -488,6 +535,8 @@ class PDONetworkManager:
         Args:
             network_type: Network type (EtherCAT or CANopen) on which to start the PDO exchange.
             refresh_rate: Determines how often (seconds) the PDO values will be updated.
+            watchdog_timeout: The PDO watchdog time. If not provided it will be set proportional
+             to the refresh rate.
 
         Raises:
             ValueError: If the refresh rate is too high.
@@ -537,6 +586,7 @@ class PDONetworkManager:
         self._pdo_thread = self.ProcessDataThread(
             net,
             refresh_rate,
+            watchdog_timeout,
             self._notify_send_process_data,
             self._notify_receive_process_data,
             self._notify_exceptions,
@@ -620,6 +670,7 @@ class PDONetworkManager:
         servo: str = DEFAULT_SERVO,
         sampling_time: float = 0.125,
         buffer_size: int = 100,
+        watchdog_timeout: Optional[float] = None,
         start: bool = True,
     ) -> PDOPoller:
         """
@@ -647,6 +698,8 @@ class PDONetworkManager:
             servo: servo alias to reference it. ``default`` by default.
             sampling_time: period of the sampling in seconds.
                 By default ``0.125`` seconds.
+            watchdog_timeout: The PDO watchdog time. If not provided it will be set proportional
+             to the refresh rate.
             buffer_size: number maximum of sample for each data read.
                 ``100`` by default.
             start: if ``True``, function starts poller, if ``False``
@@ -656,7 +709,7 @@ class PDONetworkManager:
             The poller instance.
 
         """
-        poller = PDOPoller(self.mc, servo, sampling_time, buffer_size)
+        poller = PDOPoller(self.mc, servo, sampling_time, watchdog_timeout, buffer_size)
         poller.add_channels(registers)
         if start:
             poller.start()