+++ tests/test_examples.py
@@ -1,7 +1,13 @@
+from typing import Dict
+
+import pytest
 import pytest
 from ingenialink import CAN_BAUDRATE, CAN_DEVICE
 from ingenialink.exceptions import ILFirmwareLoadError
 
+from examples.change_baudrate import change_baudrate
+from examples.change_node_id import change_node_id
+from examples.connect_ecat_coe import connect_ethercat_coe
 from examples.load_fw_canopen import load_firmware_canopen
 from examples.load_save_configuration import main as main_load_save_configuration
 from examples.load_save_config_register_changes import (
@@ -11,6 +17,7 @@ from ingeniamotion import MotionController
 from ingeniamotion.communication import Communication
 from ingeniamotion.configuration import Configuration
 from ingeniamotion.enums import SeverityLevel
+from ingeniamotion.information import Information
 
 
 @pytest.mark.eoe
@@ -22,8 +29,8 @@ def test_disturbance_example(read_config, script_runner):
     assert result.returncode == 0
 
 
+@pytest.mark.usefixtures("setup_for_test_examples", "teardown_for_test_examples")
 @pytest.mark.canopen
-@pytest.mark.skip(reason="This test fails because the canopen node is already connected")
 def test_canopen_example(read_config, script_runner):
     script_path = "examples/canopen_example.py"
     dictionary = read_config["dictionary"]
@@ -307,6 +314,222 @@ def test_can_bootloader_example_failed(mocker, capsys):
     assert all_outputs[3] == "Starts to load the firmware."
     assert all_outputs[4] == f"Firmware loading failed: {fw_error_message}"
     assert all_outputs[5] == "Drive is disconnected."
+
+
+@pytest.mark.virtual
+def test_change_node_id_success(mocker, capsys):
+    device = CAN_DEVICE.PCAN
+    channel = 0
+    baudrate = CAN_BAUDRATE.Baudrate_1M
+    dictionary_path = "test_dictionary.xdf"
+
+    node_id = 20
+    test_new_node_id = 32
+
+    mocker.patch.object(Communication, "connect_servo_canopen")
+    mocker.patch.object(Communication, "disconnect")
+    mocker.patch.object(Information, "get_node_id", side_effect=[node_id, test_new_node_id])
+    mocker.patch.object(Configuration, "change_node_id")
+    change_node_id(device, channel, node_id, baudrate, dictionary_path, test_new_node_id)
+
+    captured_outputs = capsys.readouterr()
+    all_outputs = captured_outputs.out.split("\n")
+    assert all_outputs[1] == f"Drive is connected with {node_id} as a node ID."
+    assert all_outputs[3] == "Node ID has been changed"
+    assert all_outputs[6] == f"Now the drive is connected with {test_new_node_id} as a node ID."
+
+
+@pytest.mark.virtual
+def test_change_node_id_failed(mocker, capsys):
+    device = CAN_DEVICE.PCAN
+    channel = 0
+    baudrate = CAN_BAUDRATE.Baudrate_1M
+    dictionary_path = "test_dictionary.xdf"
+
+    node_id = 20
+    test_new_node_id = node_id
+
+    mocker.patch.object(Communication, "connect_servo_canopen")
+    mocker.patch.object(Communication, "disconnect")
+    mocker.patch.object(Information, "get_node_id", side_effect=[node_id, test_new_node_id])
+    mocker.patch.object(Configuration, "change_node_id")
+    change_node_id(device, channel, node_id, baudrate, dictionary_path, test_new_node_id)
+
+    captured_outputs = capsys.readouterr()
+    all_outputs = captured_outputs.out.split("\n")
+    assert all_outputs[1] == f"Drive is connected with {node_id} as a node ID."
+    assert all_outputs[3] == f"This drive already has this node ID: {node_id}."
+
+
+@pytest.mark.virtual
+def test_change_baudrate_success(mocker, capsys):
+    device = CAN_DEVICE.PCAN
+    channel = 0
+    baudrate = CAN_BAUDRATE.Baudrate_1M
+    node_id = 32
+    dictionary_path = "test_dictionary.xdf"
+    test_new_baudrate = CAN_BAUDRATE.Baudrate_125K
+
+    mocker.patch.object(Communication, "connect_servo_canopen")
+    mocker.patch.object(Communication, "disconnect")
+    mocker.patch.object(Information, "get_baudrate", side_effect=[baudrate])
+    mocker.patch.object(Configuration, "change_baudrate")
+    change_baudrate(device, channel, node_id, baudrate, dictionary_path, test_new_baudrate)
+
+    captured_outputs = capsys.readouterr()
+    all_outputs = captured_outputs.out.split("\n")
+    assert all_outputs[0] == f"Drive is connected with {baudrate} baudrate."
+    assert all_outputs[2] == f"Baudrate has been changed from {baudrate} to {test_new_baudrate}."
+    assert (
+        all_outputs[4]
+        == f"Perform a power cycle and reconnect to the drive using the new baud rate: {test_new_baudrate}"
+    )
+
+
+@pytest.mark.virtual
+def test_change_baudrate_failed(mocker, capsys):
+    device = CAN_DEVICE.PCAN
+    channel = 0
+    baudrate = CAN_BAUDRATE.Baudrate_1M
+    node_id = 32
+    dictionary_path = "test_dictionary.xdf"
+    test_new_baudrate = baudrate
+
+    mocker.patch.object(Communication, "connect_servo_canopen")
+    mocker.patch.object(Communication, "disconnect")
+    mocker.patch.object(Information, "get_baudrate", side_effect=[baudrate])
+    mocker.patch.object(Configuration, "change_baudrate")
+    change_baudrate(device, channel, node_id, baudrate, dictionary_path, test_new_baudrate)
+
+    captured_outputs = capsys.readouterr()
+    all_outputs = captured_outputs.out.split("\n")
+    assert all_outputs[0] == f"Drive is connected with {baudrate} baudrate."
+    assert all_outputs[2] == f"This drive already has this baudrate: {baudrate}."
+
+
+@pytest.mark.virtual
+def test_ecat_coe_connection_example_success(mocker, capsys):
+    interface_index = 2
+    slave_id = 1
+    dictionary_path = (
+        "\\\\awe-srv-max-prd\\distext\\products\\CAP-NET\\firmware\\2.4.0\\cap-net-e_eoe_2.4.0.xdf"
+    )
+    expected_slave_list = [slave_id]
+    expected_interfaces_name_list = ["Interface 1", "Interface 2", "Interface 3"]
+    expected_real_name_interface = f"\\Device\\NPF_real_name_interface_{interface_index}"
+    test_alias = "default"
+    test_servos: Dict[str, str] = {}
+
+    def scan_servos_ethercat(*args, **kwargs):
+        return expected_slave_list
+
+    def get_interface_name_list(*args, **kwargs):
+        return expected_interfaces_name_list
+
+    def get_ifname_by_index(*args, **kwargs):
+        return expected_real_name_interface
+
+    def connect_servo_ethercat(*args, **kwargs):
+        test_servos[test_alias] = "my_servo"
+
+    def disconnect(*args, **kwargs):
+        test_servos.pop(test_alias)
+
+    mocker.patch.object(Communication, "scan_servos_ethercat", scan_servos_ethercat)
+    mocker.patch.object(Communication, "get_interface_name_list", get_interface_name_list)
+    mocker.patch.object(Communication, "get_ifname_by_index", get_ifname_by_index)
+    mocker.patch.object(MotionController, "servos", test_servos)
+    mocker.patch.object(Communication, "connect_servo_ethercat", connect_servo_ethercat)
+    mocker.patch.object(Communication, "disconnect", disconnect)
+    connect_ethercat_coe(interface_index, slave_id, dictionary_path)
+    captured_outputs = capsys.readouterr()
+    all_outputs = captured_outputs.out.split("\n")
+    assert all_outputs[0] == "List of interfaces:"
+    assert all_outputs[1] == "0: Interface 1"
+    assert all_outputs[2] == "1: Interface 2"
+    assert all_outputs[3] == "2: Interface 3"
+    assert all_outputs[4] == "Interface selected:"
+    assert all_outputs[5] == f"- Index interface: {interface_index}"
+    assert all_outputs[6] == f"- Interface identifier: {expected_real_name_interface}"
+    assert all_outputs[7] == f"- Interface name: {expected_interfaces_name_list[interface_index]}"
+    assert all_outputs[8] == f"Found slaves: {expected_slave_list}"
+    assert all_outputs[9] == f"Drive is connected."
+    assert all_outputs[10] == f"The drive has been disconnected."
+
+
+@pytest.mark.virtual
+def test_ecat_coe_connection_example_failed(mocker, capsys):
+    interface_index = 2
+    slave_id = 1
+    dictionary_path = (
+        "\\\\awe-srv-max-prd\\distext\\products\\CAP-NET\\firmware\\2.4.0\\cap-net-e_eoe_2.4.0.xdf"
+    )
+    expected_slave_list = []
+    expected_interfaces_name_list = ["Interface 1", "Interface 2", "Interface 3"]
+    expected_real_name_interface = f"\\Device\\NPF_real_name_interface_{interface_index}"
+    test_servos: Dict[str, str] = {}
+
+    def scan_servos_ethercat(*args, **kwargs):
+        return expected_slave_list
+
+    def get_interface_name_list(*args, **kwargs):
+        return expected_interfaces_name_list
+
+    def get_ifname_by_index(*args, **kwargs):
+        return expected_real_name_interface
+
+    mocker.patch.object(Communication, "scan_servos_ethercat", scan_servos_ethercat)
+    mocker.patch.object(Communication, "get_interface_name_list", get_interface_name_list)
+    mocker.patch.object(Communication, "get_ifname_by_index", get_ifname_by_index)
+    mocker.patch.object(MotionController, "servos", test_servos)
+    connect_ethercat_coe(interface_index, slave_id, dictionary_path)
+    captured_outputs = capsys.readouterr()
+    all_outputs = captured_outputs.out.split("\n")
+    assert all_outputs[0] == "List of interfaces:"
+    assert all_outputs[1] == "0: Interface 1"
+    assert all_outputs[2] == "1: Interface 2"
+    assert all_outputs[3] == "2: Interface 3"
+    assert all_outputs[4] == "Interface selected:"
+    assert all_outputs[5] == f"- Index interface: {interface_index}"
+    assert all_outputs[6] == f"- Interface identifier: {expected_real_name_interface}"
+    assert all_outputs[7] == f"- Interface name: {expected_interfaces_name_list[interface_index]}"
+    assert (
+        all_outputs[8]
+        == f"No slave detected on interface: {expected_interfaces_name_list[interface_index]}"
+    )
+
+
+@pytest.mark.virtual
+def test_ecat_coe_connection_example_connection_error(mocker, capsys):
+    interface_index = 2
+    slave_id = 1
+    dictionary_path = (
+        "\\\\awe-srv-max-prd\\distext\\products\\CAP-NET\\firmware\\2.4.0\\cap-net-e_eoe_2.4.0.xdf"
+    )
+    expected_interfaces_name_list = ["Interface 1", "Interface 2", "Interface 3"]
+    expected_real_name_interface = f"\\Device\\NPF_real_name_interface_{interface_index}"
+
+    def get_interface_name_list(*args, **kwargs):
+        return expected_interfaces_name_list
+
+    def get_ifname_by_index(*args, **kwargs):
+        return expected_real_name_interface
+
+    mocker.patch.object(Communication, "get_interface_name_list", get_interface_name_list)
+    mocker.patch.object(Communication, "get_ifname_by_index", get_ifname_by_index)
+    with pytest.raises(ConnectionError) as e:
+        connect_ethercat_coe(interface_index, slave_id, dictionary_path)
+    captured_outputs = capsys.readouterr()
+    all_outputs = captured_outputs.out.split("\n")
+    assert all_outputs[0] == "List of interfaces:"
+    assert all_outputs[1] == "0: Interface 1"
+    assert all_outputs[2] == "1: Interface 2"
+    assert all_outputs[3] == "2: Interface 3"
+    assert all_outputs[4] == "Interface selected:"
+    assert all_outputs[5] == f"- Index interface: {interface_index}"
+    assert all_outputs[6] == f"- Interface identifier: {expected_real_name_interface}"
+    assert all_outputs[7] == f"- Interface name: {expected_interfaces_name_list[interface_index]}"
+    assert e.value.args[0] == f"could not open interface {expected_real_name_interface}"
 
 
 @pytest.mark.virtual